+++ src/internal-apis.js
@@ -1,4 +1,4 @@
-export * from './commands/listCommits'
+export * from './commands/listCommitsAndTags'
 export * from './commands/listObjects'
 export * from './commands/pack'
 export * from './commands/uploadPack'
